# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PET - Portable Event Telephone API
  version: 1.0.0
  description: REST API for events, extensions, devices, DECT, callbacks and more.
paths:
...
  /api/v1/breakout/permissions/:
    get:
      operationId: v1_breakout_permissions_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPermissionList'
          description: ''
    post:
      operationId: v1_breakout_permissions_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PermissionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PermissionRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
          description: ''
...
  /api/v1/breakout/rules/:
    get:
      operationId: v1_breakout_rules_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRuleList'
          description: ''
    post:
      operationId: v1_breakout_rules_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RuleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RuleRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
          description: ''
...
  /api/v1/breakout/trunks/:
    get:
      operationId: v1_breakout_trunks_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTrunkList'
          description: ''
    post:
      operationId: v1_breakout_trunks_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrunkRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TrunkRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TrunkRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trunk'
          description: ''
...
  /api/v1/breakout/usage/:
    get:
      operationId: v1_breakout_usage_retrieve
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    PaginatedPermissionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
...
    PaginatedRuleList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
...
    PaginatedTrunkList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Trunk'
...
    Permission:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
        extension:
          type: string
          format: uuid
          nullable: true
        user_group:
          type: integer
          nullable: true
        allowed:
          type: boolean
        daily_minutes_limit:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          description: 0 = unlimited.
      required:
      - event
      - id
    PermissionRequest:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        extension:
          type: string
          format: uuid
          nullable: true
        user_group:
          type: integer
          nullable: true
        allowed:
          type: boolean
        daily_minutes_limit:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          description: 0 = unlimited.
      required:
      - event
...
    Rule:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        trunk:
          type: integer
        name:
          type: string
          maxLength: 80
        pattern:
          type: string
          description: Regex matched against the number after the prefix.
          maxLength: 120
        allow:
          type: boolean
        per_call_max_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          description: 0 = unlimited.
        priority:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          description: Lower wins.
      required:
      - id
      - name
      - pattern
      - trunk
    RuleRequest:
      type: object
      properties:
        trunk:
          type: integer
        name:
          type: string
          minLength: 1
          maxLength: 80
        pattern:
          type: string
          minLength: 1
          description: Regex matched against the number after the prefix.
          maxLength: 120
        allow:
          type: boolean
        per_call_max_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          description: 0 = unlimited.
        priority:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          description: Lower wins.
      required:
      - name
      - pattern
      - trunk
...
    Trunk:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
        name:
          type: string
          maxLength: 80
        sip_host:
          type: string
          maxLength: 200
        port:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        transport:
          $ref: '#/components/schemas/TrunkTransportEnum'
        auth_user:
          type: string
          maxLength: 80
        from_domain:
          type: string
          maxLength: 200
        outbound_prefix:
          type: string
          description: Digits users dial first, e.g. 0.
          maxLength: 8
        caller_id_default:
          type: string
          description: E.164 number presented by default.
          maxLength: 32
        enabled:
          type: boolean
      required:
      - event
      - id
      - name
      - sip_host
    TrunkRequest:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
        sip_host:
          type: string
          minLength: 1
          maxLength: 200
        port:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        transport:
          $ref: '#/components/schemas/TrunkTransportEnum'
        auth_user:
          type: string
          maxLength: 80
        auth_password:
          type: string
          writeOnly: true
          maxLength: 128
        from_domain:
          type: string
          maxLength: 200
        outbound_prefix:
          type: string
          minLength: 1
          description: Digits users dial first, e.g. 0.
          maxLength: 8
        caller_id_default:
          type: string
          description: E.164 number presented by default.
          maxLength: 32
        enabled:
          type: boolean
      required:
      - event
      - name
      - sip_host
    TrunkTransportEnum:
      enum:
      - udp
      - tcp
      - tls
      type: string
      description: |-
        * `udp` - UDP
        * `tcp` - TCP
        * `tls` - TLS
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid